2_basic_reflection_system/requirements.txt
langchain_core==0.3.74
langchain_google_genai==2.1.9
langgraph==0.6.4
python-dotenv==1.1.1
